Matrix-of-Truth-main/Matrix-of-Truth-main/backend_matrix/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
audioread==3.0.1
beautifulsoup4==4.13.3
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
cssselect==1.2.0
decorator==5.2.1
distro==1.9.0
fastapi==0.115.8
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.17.0
firebase-admin==6.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-cloud-vision==3.8.1
google-crc32c==1.6.0
google-generativeai==0.8.4
google-genai==0.2.2
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.44.0
lxml==5.3.1
lxml_html_clean==0.4.1
msgpack==1.1.0
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
newsapi==0.1.1
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
proto-plus==1.26.0
protobuf==5.29.3
pusher==3.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.15.2
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
starlette==0.45.3
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.1.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets>=13.0,<15.0
xgboost==2.0.3
plotly>=5.14.0
kaleido>=0.2.1
torch>=2.9.0
transformers>=4.57.0
huggingface_hub>=0.36.0
spacy>=3.5.0
tensorflow>=2.12.0
opencv-python>=4.7.0
networkx>=2.8.0
huggingface_hub>=0.16.0
sentencepiece>=0.1.95
matplotlib
neo4j
gunicorn==21.2.0
